Linear Search Algorithm
Completed the linear search algorithm. Added a do while loop so that user may keep trying to guess if the number they are looking for is in the array. Simple algorithm that goes through every single  element starting from the beginning until all the elements are completed or until the value they are looking for is found.